add generic csv import comp
